+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - In FileMetrics report, space and collection IDs are only added to the report once to avoid repeating.
 - Apply 'max' restriction when fetching dataset file lists earlier, to avoid long load times for certain previewers.
   [CATS-899](https://opensource.ncsa.illinois.edu/jira/browse/CATS-899)
+- Unable to edit metadata definition when description included newlines characters.
   
 ### Changed 
 - Extraction events on File and Dataset pages are now grouped by extractor. The events view has been moved to a tab for both,